feat: add basic templating support
There are usually dynamic parts in a config.yaml (eg. a serial for a
image). This commit adds support for templating (based on the python
string.Template class). So in the config file, identifiers can be
set (starting with a $ sign, see [0]) and those identifiers will be
replaced with a value given in a config mapping file.
That mapping file is a yaml file with a dict like structure.
So if the config.yaml looks like:

---
images:
  my-first-image-$serial

and the config-mapping.yaml looks like:
---
serial: 123

the the resulting config used for publication is:
---
images:
  my-first-image-123

[0] https://docs.python.org/3/library/string.html#string.Template