feat: use multipart s3 upload to be able to retry
Currently the used s3client.upload_file() method does already use a
multipart upload. But there's no way to retry an interrupted upload.

This change does add a multipart upload mechanism that can be
interrupted and continued in case of failures (eg. if the upload to a
S3 bucket in China fails due to network problems).

Note: it's highly recommended to setup a bucket lifecycle rule to
delete incomplete multipart uploads after some days[0]. So retries can
be done within those days but afterwards a new upload has to be
started.

Refs:
- [0]
https://docs.aws.amazon.com/AmazonS3/latest/userguide//mpu-abort-incomplete-mpu-lifecycle-config.html