feat: raise exception if s3 bucket in unexpected region
checking for the existence of the specified upload bucket does not
verify that it is in the expected region. This allows the upload to
succeed, but will fail on snapshot creation. In an effort to fail sooner
and more clearly identify issues, this change will raise an exception if
the listed bucket is in an unexpected region.