feat: remove bucket_region from config
specifying bucket_region in config makes sharing a single config across
aws partitions challenging, and also introduces the possibility of
specifying a bucket_name that exists outside of the bucket_region.
Instead of requiring users to specify the region, we instead just take
the bucket name (which is unique across all regions of a partition) and,
if the bucket exists, query aws for the region it is in.
If the bucket does not exist, raise an error indicating same.