feat: apply marketplace configuration only in default aws partition
If awspub configuration files are used in different partitions than
the default (aws) one, it doesn't make sense (and doesn't work) to try
to create a Marketplace version update. So make it clear that the
marketplace configuration only applies in the default partition.
Note: there is also a china marketplace but that's for now not
supported.