refactor: Move logic for getting regions from the config to `common.py`
Currently, the `region` key can be configured in the image
configuration. However, sibce `sns` also support the `region`
key to send a notifications to specific regions, this logic
can be resued to get the sns regions.

Move the logic to common.py and rename to `_get_regions` as
a more general function for retrieving regions from
different keys.